Updated Gradle Version Configurations
The gradle set up for this project was different from the one installed on my system, so I added code to specify which gradle version was required. This should prompt you to install the 2.3.3 version if it is not installed already. This should be a fix, but let me know if this causes problems for anyone!