Added (Possible) Swipe Detection Code
Reuben, here's the code I found that I thought would work really well for us detecting left/right/up/down swipe input from the user. I worked on implementing it but couldn't get it to work with the current view. If this helps you that's great, if not feel free to replace it with a different method!