Working on changing reference of layout
Originally the layout was built to be in reference from the player's position down. I made changes so the busses and boats are now in reference to where the player is up, so louie travels up as he moves forward in the level. I added  some changes to the background too, but it's not lining up correctly